remove ipns
This commit was moved from ipfs/go-delegated-routing@21389496b59f6a072126bf30720adbcc7819865b
